chore: move find-plugin & runtime in lib
