+++ lib/find-plugins.js
